core: implement legacy unsubscription
This changes callback_list so that all callbacks are erased if
subscribe(nullptr) is used as it was previously done.

This means that if you pass nullptr it will remove all (other)
callbacks. In consequence, software written against the old behaviour
should keep working, however, will print an error that the new API ought
to be used.

I could not find a simple way to catch subscribe(nullptr) at compile
time, unfortunately.